requirements.txt
fastapi
uvicorn
langchain
openai
chromadb
streamlit
pydantic
jinja2
faiss-cpu
python-dotenv
icalendar
